still need to convert all commands to ts

+++ src/commands/admin/delete.ts
@@ -1,4 +1,4 @@
-import { SlashCommandBuilder } from "discord.js";
+import { ChatInputCommandInteraction, SlashCommandBuilder } from "discord.js";
 import { Profile } from "../../schemas/profile.js";
 
 export default {
@@ -6,11 +6,15 @@ export default {
     .setName('delete')
     .setDescription('delete profile'),
 
-  async execute(interaction) {
+  async execute(interaction: ChatInputCommandInteraction) {
     const { user, guild } = interaction;
 
     if (user.id !== '449357416287567873')
       return interaction.reply('not an admin can not use');
+
+    if (guild == null) {
+      return interaction.reply('user does not exist');
+    }
 
     const result = await Profile.deleteOne({
       userId: user.id,
